LayoutTests/storage/indexeddb/modern/idbobjectstore-get-failures-expected.txt

ALERT: Initial upgrade needed: Old version - 0 New version - 1
ALERT: Failed to get record from object store with an invalid key
ALERT: Failed to get record from object store that has been deleted
ALERT: Initial upgrade versionchange transaction complete
ALERT: Failed to get record from object store while the transaction is inactive
ALERT: readonly transaction complete
ALERT: Done
This tests some obvious failures that can happen while calling IDBObjectStore.get().
